Move Storage.java into a package
